gmail-api/requirements.txt
autopep8==1.5.2
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-api-core==1.17.0
google-api-python-client==1.9.1
google-auth==1.16.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
tqdm==4.46.0
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
